Use views instead of DTO on client (rotation screen save is broken)

+++ optashift-employee-rostering-server/src/main/java/org/optaplanner/openshift/employeerostering/server/shift/ShiftRestServiceImpl.java
@@ -16,10 +16,13 @@
 
 package org.optaplanner.openshift.employeerostering.server.shift;
 
-import java.util.Collection;
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
+import javax.inject.Inject;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
@@ -27,16 +30,28 @@ import javax.transaction.Transactional;
 
 import org.optaplanner.openshift.employeerostering.server.common.AbstractRestServiceImpl;
 import org.optaplanner.openshift.employeerostering.shared.employee.Employee;
+import org.optaplanner.openshift.employeerostering.shared.employee.EmployeeRestService;
+import org.optaplanner.openshift.employeerostering.shared.roster.RosterState;
 import org.optaplanner.openshift.employeerostering.shared.rotation.ShiftTemplate;
+import org.optaplanner.openshift.employeerostering.shared.rotation.view.RotationView;
+import org.optaplanner.openshift.employeerostering.shared.rotation.view.ShiftTemplateView;
 import org.optaplanner.openshift.employeerostering.shared.shift.Shift;
 import org.optaplanner.openshift.employeerostering.shared.shift.ShiftRestService;
 import org.optaplanner.openshift.employeerostering.shared.shift.view.ShiftView;
 import org.optaplanner.openshift.employeerostering.shared.spot.Spot;
+import org.optaplanner.openshift.employeerostering.shared.spot.SpotRestService;
+import org.optaplanner.openshift.employeerostering.shared.tenant.TenantRestService;
 
 public class ShiftRestServiceImpl extends AbstractRestServiceImpl implements ShiftRestService {
 
     @PersistenceContext
     private EntityManager entityManager;
+    @Inject
+    private TenantRestService tenantRestService;
+    @Inject
+    private SpotRestService spotRestService;
+    @Inject
+    private EmployeeRestService employeeRestService;
 
     @Override
     @Transactional
@@ -76,7 +91,7 @@ public class ShiftRestServiceImpl extends AbstractRestServiceImpl implements Shi
             validateTenantIdParameter(tenantId, rotationEmployee);
         }
 
-        Shift shift = new Shift(shiftView, spot, shiftView.getStartDateTime(), shiftView.getEndDateTime(),
+        Shift shift = new Shift(tenantRestService.getTenantConfiguration(tenantId).getTimeZone(), shiftView, spot,
                 rotationEmployee);
         shift.setPinnedByUser(shiftView.isPinnedByUser());
         Long employeeId = shiftView.getEmployeeId();
@@ -115,18 +130,60 @@ public class ShiftRestServiceImpl extends AbstractRestServiceImpl implements Shi
     }
 
     @Override
-    public List<ShiftTemplate> getShiftTemplateList(Integer tenantId) {
-        TypedQuery<ShiftTemplate> q = entityManager.createNamedQuery("ShiftTemplate.findAll", ShiftTemplate.class);
-        q.setParameter("tenantId", tenantId);
-        return q.getResultList();
+    public RotationView getRotation(Integer tenantId) {
+        List<ShiftTemplate> shiftTemplateList = entityManager.createNamedQuery("ShiftTemplate.findAll", ShiftTemplate.class)
+                .setParameter("tenantId", tenantId)
+                .getResultList();
+        RotationView out = new RotationView();
+        out.setTenantId(tenantId);
+        out.setSpotList(spotRestService.getSpotList(tenantId));
+        out.setEmployeeList(employeeRestService.getEmployeeList(tenantId));
+        out.setRotationLength(entityManager.createNamedQuery("RosterState.find", RosterState.class)
+                .setParameter("tenantId", tenantId)
+                .getSingleResult().getRotationLength());
+        Map<Long, List<ShiftTemplateView>> spotIdToShiftTemplateViewListMap = new HashMap<>();
+        shiftTemplateList.forEach((shiftTemplate) -> {
+            spotIdToShiftTemplateViewListMap.computeIfAbsent(shiftTemplate.getSpot().getId(),
+                    (k) -> new ArrayList<>())
+                    .add(new ShiftTemplateView(out.getRotationLength(), shiftTemplate));
+        });
+        out.setSpotIdToShiftTemplateViewListMap(spotIdToShiftTemplateViewListMap);
+        return out;
     }
 
     @Override
     @Transactional
-    public void updateShiftTemplate(Integer tenantId, List<ShiftTemplate> shiftTemplateList) {
-        List<ShiftTemplate> oldShiftTemplateList = getShiftTemplateList(tenantId);
+    public void updateRotation(Integer tenantId, RotationView rotationView) {
+        if (!tenantId.equals(rotationView.getTenantId())) {
+            throw new IllegalArgumentException("rotationView (" + rotationView + ") tenantId" +
+                    " does not match tenantId (" + tenantId + ")");
+        }
+        List<ShiftTemplate> oldShiftTemplateList = entityManager.createNamedQuery("ShiftTemplate.findAll", ShiftTemplate.class)
+                .setParameter("tenantId", tenantId)
+                .getResultList();
         oldShiftTemplateList.forEach((s) -> entityManager.remove(s));
-        shiftTemplateList.forEach((s) -> entityManager.persist(s));
+        // TODO: Update rotation length and unplanneOffset
+        Integer rotationLength = entityManager.createNamedQuery("RosterState.find", RosterState.class)
+                .setParameter("tenantId", tenantId)
+                .getSingleResult().getRotationLength();
+        Map<Long, Spot> spotIdToSpotMap = spotRestService
+                .getSpotList(tenantId).stream().collect(Collectors
+                        .toMap(spot -> spot.getId(), spot -> spot));
+        Map<Long, Employee> employeeIdToEmployeeMap = employeeRestService
+                .getEmployeeList(tenantId).stream().collect(Collectors
+                        .toMap(employee -> employee.getId(), employee -> employee));
+        rotationView.getSpotIdToShiftTemplateViewListMap()
+                .forEach((spotId, shiftTemplateViewList) -> {
+                    Spot spot = spotIdToSpotMap.get(spotId);
+                    if (shiftTemplateViewList != null) {
+                        shiftTemplateViewList.forEach(shiftTemplateView -> {
+                            entityManager.merge(new ShiftTemplate(rotationLength,
+                                    shiftTemplateView, spot,
+                                    employeeIdToEmployeeMap.get(shiftTemplateView.getId())));
+                        });
+                    }
+                });
+
     }
 
 }
